添加五档行情 K线功能按钮放大 日线回调NULL时崩溃BUG，当为NULL时判断不做数据换算

+++ CultureTrade/Classes/Trade/DataModel.h
@@ -43,7 +43,7 @@
 @end
 
 @interface VolPriceNS : NSObject
-@property (nonatomic,assign) int		m_uiPrice;//价格
+@property (nonatomic,assign) float		m_uiPrice;//价格
 @property (nonatomic,assign) int		m_uiVolume;//量
 @end
